snapshotsupport abstract
