+++ service/src/main/java/org/databiosphere/workspacedataservice/dataimport/snapshotsupport/RawlsSnapshotSupport.java
@@ -1,24 +1,22 @@
 package org.databiosphere.workspacedataservice.dataimport.snapshotsupport;
 
-import bio.terra.datarepo.model.RelationshipModel;
-import bio.terra.datarepo.model.TableModel;
-import com.google.common.collect.Multimap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import bio.terra.workspace.model.ResourceList;
 import java.util.UUID;
 import org.databiosphere.workspacedataservice.activitylog.ActivityLogger;
 import org.databiosphere.workspacedataservice.rawls.RawlsClient;
 import org.databiosphere.workspacedataservice.retry.RestClientRetry;
-import org.databiosphere.workspacedataservice.shared.model.RecordType;
 import org.databiosphere.workspacedataservice.shared.model.WorkspaceId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-public class RawlsSnapshotSupport implements SnapshotSupport {
+public class RawlsSnapshotSupport extends SnapshotSupport {
 
   private final WorkspaceId workspaceId;
   private final RawlsClient rawlsClient;
   private final RestClientRetry restClientRetry;
   private final ActivityLogger activityLogger;
+
+  private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
   public RawlsSnapshotSupport(
       WorkspaceId workspaceId,
@@ -31,27 +29,18 @@ public class RawlsSnapshotSupport implements SnapshotSupport {
     this.restClientRetry = restClientRetry;
   }
 
-  @Override
-  public List<UUID> existingPolicySnapshotIds(int pageSize) {
-    // TODO: implement
-    return null;
+  protected void linkSnapshot(UUID snapshotId) {
+    RestClientRetry.VoidRestCall voidRestCall =
+        (() -> rawlsClient.createSnapshotReference(workspaceId.id(), snapshotId));
+    restClientRetry.withRetryAndErrorHandling(voidRestCall, "Rawls.createSnapshotReference");
+    activityLogger.saveEventForCurrentUser(
+        user -> user.linked().snapshotReference().withUuid(snapshotId));
   }
 
-  @Override
-  public void linkSnapshots(Set<UUID> snapshotIds) {
-    // TODO: implement
-  }
-
-  @Override
-  public Map<RecordType, String> identifyPrimaryKeys(List<TableModel> tables) {
-    // TODO implement or determine whether this belongs here
-    return null;
-  }
-
-  @Override
-  public Multimap<RecordType, RelationshipModel> identifyRelations(
-      List<RelationshipModel> relationshipModels) {
-    // TODO implement
-    return null;
+  protected ResourceList enumerateDataRepoSnapshotReferences(int offset, int pageSize) {
+    RestClientRetry.RestCall<ResourceList> restCall =
+        (() -> rawlsClient.enumerateDataRepoSnapshotReferences(workspaceId.id(), offset, pageSize));
+    return restClientRetry.withRetryAndErrorHandling(
+        restCall, "Rawls.enumerateDataRepoSnapshotReferences");
   }
 }
